refactor:修正將驗證規則留在子組件
