+++ src/views/RegisterPage.vue
@@ -10,7 +10,7 @@ import TheSvg from '@/components/global/TheSvg.vue'
 const authStore = useAuthStore()
 const isLoading = ref<boolean>(false)
 const router: Router = useRouter()
-const ruleFormRef = ref<FormInstance>()
+// const ruleFormRef = ref<FormInstance>()
 type RegisterInputType = {
   account: string
   password: string
@@ -45,57 +45,57 @@ const authButtonData = [
     }
   }
 ]
-const validatePass = (rule: any, value: any, callback: any) => {
-  if (value === '') {
-    callback(new Error('請輸入密碼'))
-  } else {
-    if (registerInput.value.password !== '') {
-      if (!ruleFormRef.value) return
-      ruleFormRef.value.validateField('checkPassWord')
-    }
-    callback()
-  }
-}
-const validatePass2 = (rule: any, value: any, callback: any) => {
-  if (value === '') {
-    callback(new Error('請輸入密碼'))
-  } else if (value !== registerInput.value.password) {
-    callback(new Error('密碼與原先不符合'))
-  } else {
-    callback()
-  }
-}
+// const validatePass = (rule: any, value: any, callback: any) => {
+//   if (value === '') {
+//     callback(new Error('請輸入密碼'))
+//   } else {
+//     if (registerInput.value.password !== '') {
+//       if (!ruleFormRef.value) return
+//       ruleFormRef.value?.validateField('checkPassWord')
+//     }
+//     callback()
+//   }
+// }
+// const validatePass2 = (rule: any, value: any, callback: any) => {
+//   if (value === '') {
+//     callback(new Error('請輸入密碼'))
+//   } else if (value !== registerInput.value.password) {
+//     callback(new Error('密碼與原先不符合'))
+//   } else {
+//     callback()
+//   }
+// }
 
-const registerRules = ref<FormRules>({
-  account: [
-    {
-      type: 'email',
-      required: true,
-      message: '信箱格式不相符',
-      trigger: ['blur', 'change']
-    }
-  ],
-  password: [
-    { min: 2, max: 30, message: '長度介於2到30之間', trigger: 'blur' },
-    { required: true, message: '必填', trigger: 'blur' },
-    { validator: validatePass, trigger: 'blur' }
-  ],
-  checkPassWord: [
-    { min: 6, max: 30, message: '長度介於2到30之間', trigger: 'blur' },
-    { required: true, message: '必填', trigger: 'blur' },
-    { validator: validatePass2, trigger: 'blur' }
-  ],
-  privacy: [
-    {
-      type: 'array',
-      required: true,
-      message: '請詳細閱讀隱私條款',
-      trigger: 'change'
-    }
-  ]
-})
+// const registerRules = ref<FormRules>({
+//   account: [
+//     {
+//       type: 'email',
+//       required: true,
+//       message: '信箱格式不相符',
+//       trigger: ['blur', 'change']
+//     }
+//   ],
+//   password: [
+//     { min: 2, max: 30, message: '長度介於2到30之間', trigger: 'blur' },
+//     { required: true, message: '必填', trigger: 'blur' },
+//     { validator: validatePass, trigger: 'blur' }
+//   ],
+//   checkPassWord: [
+//     { min: 6, max: 30, message: '長度介於2到30之間', trigger: 'blur' },
+//     { required: true, message: '必填', trigger: 'blur' },
+//     { validator: validatePass2, trigger: 'blur' }
+//   ],
+//   privacy: [
+//     {
+//       type: 'array',
+//       required: true,
+//       message: '請詳細閱讀隱私條款',
+//       trigger: 'change'
+//     }
+//   ]
+// })
 const handleRegister = async (formEl: FormInstance | undefined) => {
-  ruleFormRef.value = formEl
+  // ruleFormRef.value = formEl
   if (!formEl) return
   await formEl.validate(async (valid, fields) => {
     if (valid) {
@@ -137,10 +137,8 @@ const fetchRegister = async (data: RegisterInputType) => {
         <!-- <RegisterForm /> -->
         <RegisterForm
           v-model:registerInput="registerInput"
-          :rules="registerRules"
           :loading="isLoading"
           :handleRegister="handleRegister"
-          ref="ruleFormRef"
         />
         <el-divider><p style="color: #9cb0c9">OR</p></el-divider>
         <ExternalAuthButton
